+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes will be documented in this file.
 
 ### Added
 
+* Move cache utilities to API (#111)
 * Promoted `server::TraceInfo` to API
 * Add exchange sequence number to `MarketByPrice` and `MarketByOrder` (#101)
 * Add `max_trade_vol` and `trade_vol_step_size` to ReferenceData (#100)